+++ konva.js
@@ -5,10 +5,10 @@
 })(this, (function () { 'use strict';
 
   /*
-   * Konva JavaScript Framework v8.3.5
+   * Konva JavaScript Framework v8.3.12
    * http://konvajs.org/
    * Licensed under the MIT
-   * Date: Mon Mar 21 2022
+   * Date: Mon Aug 29 2022
    *
    * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
    * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -35,7 +35,7 @@
               : {};
   const Konva$2 = {
       _global: glob,
-      version: '8.3.5',
+      version: '8.3.12',
       isBrowser: detectBrowser(),
       isUnminified: /param/.test(function (param) { }.toString()),
       dblClickWindow: 400,
@@ -369,17 +369,6 @@
        */
       getMatrix() {
           return this.m;
-      }
-      /**
-       * set to absolute position via translation
-       * @method
-       * @name Konva.Transform#setAbsolutePosition
-       * @returns {Konva.Transform}
-       * @author ericdrowell
-       */
-      setAbsolutePosition(x, y) {
-          var m0 = this.m[0], m1 = this.m[1], m2 = this.m[2], m3 = this.m[3], m4 = this.m[4], m5 = this.m[5], yt = (m0 * (y - m5) - m1 * (x - m4)) / (m0 * m3 - m1 * m2), xt = (x - m4 - m2 * yt) / m0;
-          return this.translate(xt, yt);
       }
       /**
        * convert transformation matrix back into node's attributes
@@ -1182,6 +1171,12 @@
   function getNumberArrayValidator() {
       if (Konva$2.isUnminified) {
           return function (val, attr) {
+              // Retrieve TypedArray constructor as found in MDN (if TypedArray is available)
+              // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray#description
+              const TypedArray = Int8Array ? Object.getPrototypeOf(Int8Array) : null;
+              if (TypedArray && val instanceof TypedArray) {
+                  return val;
+              }
               if (!Util._isArray(val)) {
                   Util.warn(_formatValue(val) +
                       ' is a not valid value for "' +
@@ -1220,6 +1215,10 @@
   function getComponentValidator(components) {
       if (Konva$2.isUnminified) {
           return function (val, attr) {
+              // ignore validation on undefined value, because it will reset to defalt
+              if (val === undefined || val === null) {
+                  return val;
+              }
               if (!Util.isObject(val)) {
                   Util.warn(_formatValue(val) +
                       ' is a not valid value for "' +
@@ -1293,6 +1292,11 @@
                       continue;
                   }
                   this._setAttr(attr + capitalize(key), val[key]);
+              }
+              if (!val) {
+                  components.forEach((component) => {
+                      this._setAttr(attr + capitalize(component), undefined);
+                  });
               }
               this._fireChangeEvent(attr, oldVal, val);
               if (after) {
@@ -2353,6 +2357,7 @@
       // dragBefore and dragAfter allows us to set correct order of events
       // setup all in dragbefore, and stop dragging only after pointerup triggered.
       _endDragBefore(evt) {
+          const drawNodes = [];
           DD._dragElements.forEach((elem) => {
               const { node } = elem;
               // we need to find pointer relative to that node
@@ -2375,9 +2380,15 @@
               }
               const drawNode = elem.node.getLayer() ||
                   (elem.node instanceof Konva$2['Stage'] && elem.node);
-              if (drawNode) {
-                  drawNode.batchDraw();
+              if (drawNode && drawNodes.indexOf(drawNode) === -1) {
+                  drawNodes.push(drawNode);
               }
+          });
+          // draw in a sync way
+          // because mousemove event may trigger BEFORE batch draw is called
+          // but as we have not hit canvas updated yet, it will trigger incorrect mouseover/mouseout events
+          drawNodes.forEach((drawNode) => {
+              drawNode.draw();
           });
       },
       _endDragAfter(evt) {
@@ -3982,6 +3993,9 @@
                   (stage ? stage.height() : 0),
               pixelRatio: pixelRatio,
           }), context = canvas.getContext();
+          if (config.imageSmoothingEnabled === false) {
+              context._context.imageSmoothingEnabled = false;
+          }
           context.save();
           if (x || y) {
               context.translate(-1 * x, -1 * y);
@@ -4004,6 +4018,7 @@
        * You can use that property to increase quality of the image, for example for super hight quality exports
        * or usage on retina (or similar) displays. pixelRatio will be used to multiply the size of exported image.
        * If you export to 500x500 size with pixelRatio = 2, then produced image will have size 1000x1000.
+       * @param {Boolean} [config.imageSmoothingEnabled] set this to false if you want to disable imageSmoothing
        * @example
        * var canvas = node.toCanvas();
        */
@@ -4030,6 +4045,7 @@
        * You can use that property to increase quality of the image, for example for super hight quality exports
        * or usage on retina (or similar) displays. pixelRatio will be used to multiply the size of exported image.
        * If you export to 500x500 size with pixelRatio = 2, then produced image will have size 1000x1000.
+       * @param {Boolean} [config.imageSmoothingEnabled] set this to false if you want to disable imageSmoothing
        * @returns {String}
        */
       toDataURL(config) {
@@ -4043,12 +4059,13 @@
       }
       /**
        * converts node into an image.  Since the toImage
-       *  method is asynchronous, a callback is required.  toImage is most commonly used
+       *  method is asynchronous, the resulting image can only be retrieved from the config callback
+       *  or the returned Promise.  toImage is most commonly used
        *  to cache complex drawings as an image so that they don't have to constantly be redrawn
        * @method
        * @name Konva.Node#toImage
        * @param {Object} config
-       * @param {Function} config.callback function executed when the composite has completed
+       * @param {Function} [config.callback] function executed when the composite has completed
        * @param {String} [config.mimeType] can be "image/png" or "image/jpeg".
        *  "image/png" is the default
        * @param {Number} [config.x] x position of canvas section
@@ -4062,6 +4079,8 @@
        * You can use that property to increase quality of the image, for example for super hight quality exports
        * or usage on retina (or similar) displays. pixelRatio will be used to multiply the size of exported image.
        * If you export to 500x500 size with pixelRatio = 2, then produced image will have size 1000x1000.
+       * @param {Boolean} [config.imageSmoothingEnabled] set this to false if you want to disable imageSmoothing
+       * @return {Promise<Image>}
        * @example
        * var image = node.toImage({
        *   callback(img) {
@@ -4070,13 +4089,56 @@
        * });
        */
       toImage(config) {
-          if (!config || !config.callback) {
-              throw 'callback required for toImage method config argument';
-          }
-          var callback = config.callback;
-          delete config.callback;
-          Util._urlToImage(this.toDataURL(config), function (img) {
-              callback(img);
+          return new Promise((resolve, reject) => {
+              try {
+                  const callback = config === null || config === void 0 ? void 0 : config.callback;
+                  if (callback)
+                      delete config.callback;
+                  Util._urlToImage(this.toDataURL(config), function (img) {
+                      resolve(img);
+                      callback === null || callback === void 0 ? void 0 : callback(img);
+                  });
+              }
+              catch (err) {
+                  reject(err);
+              }
+          });
+      }
+      /**
+       * Converts node into a blob.  Since the toBlob method is asynchronous,
+       *  the resulting blob can only be retrieved from the config callback
+       *  or the returned Promise.
+       * @method
+       * @name Konva.Node#toBlob
+       * @param {Object} config
+       * @param {Function} [config.callback] function executed when the composite has completed
+       * @param {Number} [config.x] x position of canvas section
+       * @param {Number} [config.y] y position of canvas section
+       * @param {Number} [config.width] width of canvas section
+       * @param {Number} [config.height] height of canvas section
+       * @param {Number} [config.pixelRatio] pixelRatio of output canvas. Default is 1.
+       * You can use that property to increase quality of the image, for example for super hight quality exports
+       * or usage on retina (or similar) displays. pixelRatio will be used to multiply the size of exported image.
+       * If you export to 500x500 size with pixelRatio = 2, then produced image will have size 1000x1000.
+       * @param {Boolean} [config.imageSmoothingEnabled] set this to false if you want to disable imageSmoothing
+       * @example
+       * var blob = await node.toBlob({});
+       * @returns {Promise<Blob>}
+       */
+      toBlob(config) {
+          return new Promise((resolve, reject) => {
+              try {
+                  const callback = config === null || config === void 0 ? void 0 : config.callback;
+                  if (callback)
+                      delete config.callback;
+                  this.toCanvas(config).toBlob((blob) => {
+                      resolve(blob);
+                      callback === null || callback === void 0 ? void 0 : callback(blob);
+                  });
+              }
+              catch (err) {
+                  reject(err);
+              }
           });
       }
       setSize(size) {
@@ -6329,7 +6391,10 @@
           Konva$2['_' + eventType + 'ListenClick'] = false;
           // always call preventDefault for desktop events because some browsers
           // try to drag and drop the canvas element
-          if (evt.cancelable) {
+          // TODO: are we sure we need to prevent default at all?
+          // do not call this function on mobile because it prevent "click" event on all parent containers
+          // but apps may listen to it.
+          if (evt.cancelable && eventType !== 'touch') {
               evt.preventDefault();
           }
       }
@@ -11813,6 +11878,9 @@
       'text',
       'width',
       'height',
+      'pointerDirection',
+      'pointerWidth',
+      'pointerHeight',
   ], CHANGE_KONVA$1 = 'Change.konva', NONE$1 = 'none', UP = 'up', RIGHT$1 = 'right', DOWN = 'down', LEFT$1 = 'left', 
   // cached variables
   attrChangeListLen$1 = ATTR_CHANGE_LIST$2.length;
@@ -11980,7 +12048,11 @@
           let bottomLeft = 0;
           let bottomRight = 0;
           if (typeof cornerRadius === 'number') {
-              topLeft = topRight = bottomLeft = bottomRight = Math.min(cornerRadius, width / 2, height / 2);
+              topLeft =
+                  topRight =
+                      bottomLeft =
+                          bottomRight =
+                              Math.min(cornerRadius, width / 2, height / 2);
           }
           else {
               topLeft = Math.min(cornerRadius[0] || 0, width / 2, height / 2);
@@ -13535,19 +13607,9 @@
                           this._addTextLine(match);
                           textWidth = Math.max(textWidth, matchWidth);
                           currentHeightPx += lineHeightPx;
-                          if (!shouldWrap ||
-                              (fixedHeight && currentHeightPx + lineHeightPx > maxHeightPx)) {
-                              var lastLine = this.textArr[this.textArr.length - 1];
-                              if (lastLine) {
-                                  if (shouldAddEllipsis) {
-                                      var haveSpace = this._getTextWidth(lastLine.text + ELLIPSIS) < maxWidth;
-                                      if (!haveSpace) {
-                                          lastLine.text = lastLine.text.slice(0, lastLine.text.length - 3);
-                                      }
-                                      this.textArr.splice(this.textArr.length - 1, 1);
-                                      this._addTextLine(lastLine.text + ELLIPSIS);
-                                  }
-                              }
+                          var shouldHandleEllipsis = this._shouldHandleEllipsis(currentHeightPx);
+                          if (shouldHandleEllipsis) {
+                              this._tryToAddEllipsisToLastLine();
                               /*
                                * stop wrapping if wrapping is disabled or if adding
                                * one more line would overflow the fixed height
@@ -13579,6 +13641,9 @@
                   this._addTextLine(line);
                   currentHeightPx += lineHeightPx;
                   textWidth = Math.max(textWidth, lineWidth);
+                  if (this._shouldHandleEllipsis(currentHeightPx) && i < max - 1) {
+                      this._tryToAddEllipsisToLastLine();
+                  }
               }
               // if element height is fixed, abort if adding one more line would overflow
               if (fixedHeight && currentHeightPx + lineHeightPx > maxHeightPx) {
@@ -13594,6 +13659,33 @@
           //     maxTextWidth = Math.max(maxTextWidth, this.textArr[j].width);
           // }
           this.textWidth = textWidth;
+      }
+      /**
+       * whether to handle ellipsis, there are two cases:
+       * 1. the current line is the last line
+       * 2. wrap is NONE
+       * @param {Number} currentHeightPx
+       * @returns
+       */
+      _shouldHandleEllipsis(currentHeightPx) {
+          var fontSize = +this.fontSize(), lineHeightPx = this.lineHeight() * fontSize, height = this.attrs.height, fixedHeight = height !== AUTO && height !== undefined, padding = this.padding(), maxHeightPx = height - padding * 2, wrap = this.wrap(), shouldWrap = wrap !== NONE;
+          return (!shouldWrap ||
+              (fixedHeight && currentHeightPx + lineHeightPx > maxHeightPx));
+      }
+      _tryToAddEllipsisToLastLine() {
+          var width = this.attrs.width, fixedWidth = width !== AUTO && width !== undefined, padding = this.padding(), maxWidth = width - padding * 2, shouldAddEllipsis = this.ellipsis();
+          var lastLine = this.textArr[this.textArr.length - 1];
+          if (!lastLine || !shouldAddEllipsis) {
+              return;
+          }
+          if (fixedWidth) {
+              var haveSpace = this._getTextWidth(lastLine.text + ELLIPSIS) < maxWidth;
+              if (!haveSpace) {
+                  lastLine.text = lastLine.text.slice(0, lastLine.text.length - 3);
+              }
+          }
+          this.textArr.splice(this.textArr.length - 1, 1);
+          this._addTextLine(lastLine.text + ELLIPSIS);
       }
       // for text we can't disable stroke scaling
       // if we do, the result will be unexpected
@@ -14136,7 +14228,7 @@
                           pathCmd = undefined;
                       }
                   }
-                  if (pathCmd === {} || p0 === undefined) {
+                  if (Object.keys(pathCmd).length === 0 || p0 === undefined) {
                       return undefined;
                   }
                   var needNewSegment = false;
@@ -15498,6 +15590,16 @@
       // we will recreate inner nodes manually
       toObject() {
           return Node.prototype.toObject.call(this);
+      }
+      getClientRect() {
+          if (this.nodes().length > 0) {
+              return super.getClientRect();
+          }
+          else {
+              // if we are detached return zero size
+              // so it will be skipped in calculations
+              return { x: 0, y: 0, width: 0, height: 0 };
+          }
       }
   }
   function validateAnchors(val) {